Use generics in associated const CTFE timing example
This makes the example a bit bigger, but makes the reason why constants don't
immediately undergo CTFE more conspicuous. When there is computation in the
definition that use generic parameters, there is not enough information to
fully evaluate the definition eagerly.

I made it a `compile_fail` example so the run button on the page gives reader
useful outputs.