Remove confusing usage of terminology "block of a function"
It's usually referred to as the "body" of the function, and I'm making it consistent here.
In order to emphasize the nature of the body of a block, occasionally "body block" is used.
(One instance of "body block" was already present.)

This particularly improves the sentence that used to explain
"The block of a function is conceptually wrapped in a block that binds ..."
which was just plain confusing.