Follow the authoring guide in the authoring guide
Let's follow our own rules in the authoring guide by, e.g., using
`---` rather than the Unicode em-dash character, not wrapping lines,
specifying the language for code block sections, etc.

To improve formatting, we broke out some descriptions that were in
code block sections into the main text.  We replaced references to
"the spec" with "the reference" for consistency, as that's what most
of this document uses.  And we made some other small editorial
improvements.