Revise changes for `unsafe extern` blocks
We made an editing pass on the changes for `unsafe extern` blocks.  We
fixed some editorial items and aligned some stylistic points with the
rest of the Reference.  We did some rewording that also had the effect
of minimizing the changes a bit further.

In the examples, we added `safe` and `unsafe` qualifiers to items, as
the RFC suggests that this is the preferred way to write these.