Switch to automatic standard library links
This removes the explicit links to the standard library. In particular,
this makes it nicer to view locally since you can set SPEC_RELATIVE=0 to
make the links work.

There are a bunch of changes to the actual URL because rustdoc resolves
re-exports to link to the original definition instead of the re-export
site. From what I can tell, everything should otherwise be the same.

Not all links were able to be converted due to some limitations in
rustdoc, such as:

- Links to rexports from std_arch don't work due to https://github.com/rust-lang/rust/issues/96506.
- Links to keywords aren't supported.
- Links to trait impls where the trait is not in the prelude doesn't work (they must be in scope).