Allow admonition blockquotes with blank lines
This fixes an issue where if an admonition blockquote had multiple
paragraphs, it would incorrectly interpret it as multiple blockquotes,
and break the formatting. The problem is that the regex was expecting at
least one space after the `>`.

I vaguely recall doing this intentionally, but I don't remember now why
I did it this way, and I can't find any problems with this approach.